dev_tools/requirements.txt
Pillow==8.3.2
scikit-learn==0.24.2
torch==1.9.0
torchvision==0.10.0
nb-black==1.0.7
numpy==1.20.1
matplotlib==3.3.4
seaborn==0.11.1

sudo apt update
sudo apt install python3-opencv libcamera-apps python3-gpiozero
sudo apt install libcamera0

sudo apt install libcamera-apps
pip3 install opencv-python pillow
